+++ src/PRNG.cpp
@@ -41,7 +41,7 @@ void jump_nrand()
 	static const uint64_t JUMP[] = { 0xbeac0467eba5facb, 0xd86b048b86aa9922 };
 	uint64_t s0 = 0;
 	uint64_t s1 = 0;
-	for (int i = 0; i < sizeof JUMP / sizeof *JUMP; i++)
+	for (unsigned long long int i = 0; i < sizeof JUMP / sizeof *JUMP; i++)
 		for (int b = 0; b < 64; b++) {
 			if (JUMP[i] & UINT64_C(1) << b) {
 				s0 ^= s[0];